+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 3.0.0-beta.2
 
-### WIP
+### August 4, 2019
 
 #### Breaking changes
 
@@ -16,6 +16,7 @@
 
 - SAML authentication support
 - Remember selected connection id / schema toggle
+- Configurable session time and secret
 - Support for JSON and INI config file added. File should config using `key` fields found in [configItems.js](https://github.com/rickbergfalk/sqlpad/blob/master/server/lib/config/configItems.js). Config file path default is `$HOME/.sqlpadrc` and may otherwise be specified using `--config` via command line or `SQLPAD_CONFIG` environment variable.
 
 #### Fixes
@@ -25,6 +26,7 @@
 - Add loading indicator for schema sidebar
 - Fix user needing to sign in again after sign up
 - Fix app menu for non-admin users
+- Fix frozen editor after query error
 
 ## 3.0.0-beta.1